Youtube/requirement.txt
altgraph==0.17
future==0.18.2
pefile==2019.4.18
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
pytube3==9.6.4
pywin32-ctypes==0.2.0
typing-extensions==3.7.4.2
